botutil/requirements.txt
# Function dependencies, for example :
# package>=version
pysftp
google-cloud-storage==1.30.0
google-cloud-kms==1.4.0
google-cloud-logging==1.15.0